fix cast type for pre-031 instances

+++ projects/demo/models/staging/decidim/taxonomies/stg_decidim_taxonomies.sql
@@ -23,17 +23,17 @@
     FROM {{ source('decidim', 'decidim_taxonomies') }}
 {% else %}
     SELECT
-        CAST(NULL AS INTEGER) AS id,
+        CAST(NULL AS BIGINT) AS id,
         CAST(NULL AS TEXT) AS name,
-        CAST(NULL AS INTEGER) AS decidim_organization_id,
-        CAST(NULL AS INTEGER) AS parent_id,
-        CAST(NULL AS INTEGER) AS weight,
-        CAST(NULL AS INTEGER) AS children_count,
-        CAST(NULL AS INTEGER) AS taxonomizations_count,
+        CAST(NULL AS BIGINT) AS decidim_organization_id,
+        CAST(NULL AS BIGINT) AS parent_id,
+        CAST(NULL AS BIGINT) AS weight,
+        CAST(NULL AS BIGINT) AS children_count,
+        CAST(NULL AS BIGINT) AS taxonomizations_count,
         CAST(NULL AS TIMESTAMP) AS created_at,
         CAST(NULL AS TIMESTAMP) AS updated_at,
-        CAST(NULL AS INTEGER) AS filters_count,
-        CAST(NULL AS INTEGER) AS filter_items_count,
+        CAST(NULL AS BIGINT) AS filters_count,
+        CAST(NULL AS BIGINT) AS filter_items_count,
         CAST(NULL AS JSONB) AS part_of
     LIMIT 0
 {% endif %}
